% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_siteday_cdd.R
\name{add_siteday_cdd}
\alias{add_siteday_cdd}
\title{Add a cumulative degree day value for each year + year_day + site combination}
\usage{
add_siteday_cdd(data, ub = 120, lb = 50)
}
\arguments{
\item{data:}{data.frame with latitude, longitude, year, and year_day columns}

\item{ub:}{the upper bound for degree day calculation in degrees Fahrenheit}

\item{lb:}{the lower bound for degree day calculation in degrees Fahrenheit}
}
\value{
A data.frame with with a cumulative degree day added for each year + year_day + site
}
\description{
Add a cumulative degree day value for each year + year_day + site combination
}
