% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_yearday.R
\name{add_yearday}
\alias{add_yearday}
\title{Add yearday and year columns to a data.frame}
\usage{
add_yearday(data, date_col, datetime = F, keep_date_col = T)
}
\arguments{
\item{data:}{a data.frame with a date or datetime column}

\item{date_col:}{fa quoted name of the date or datetime column}

\item{datetime:}{should the date_col be parsed as a datetime?}

\item{keep_date_col:}{should the parsed date_col be kept?}
}
\value{
A data.frame with yearday and year columns
}
\description{
Add yearday and year columns to a data.frame
}
\examples{
add_yearday(data.frame(today()))
}
