% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_anova.R
\name{export_anova}
\alias{export_anova}
\title{Title}
\usage{
export_anova(
  model,
  file = NULL,
  pnames = NULL,
  alpha = 0.05,
  boldsig = TRUE,
  digits = c(1, 2, 1, 4),
  align = c("l", "r", "r", "r"),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
