% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamplot.R
\name{gamplot}
\alias{gamplot}
\title{GAM Quick Plot}
\usage{
gamplot(
  model,
  plot_x = NULL,
  ylab = NULL,
  xlab = NULL,
  ylims = NULL,
  xlims = NULL
)
}
\arguments{
\item{model}{A mgcv::gam model object}

\item{plot_x}{A column name for the axis of plot}

\item{ylab}{A user-provided string for the y-axis}

\item{xlab}{A user-provided string for x-axis}

\item{ylims}{A numeric vector of length two for the y limits}
}
\value{
Plots the raw response against the first independent variable,
then overlays a prediction ribbon.
}
\description{
GAM Quick Plot
}
