% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daymet.R
\name{get_daymet}
\alias{get_daymet}
\title{Get Daymet data for a given site and set of years}
\usage{
get_daymet(
  site = "Daymet",
  lat = 36.0133,
  lon = -84.2625,
  start = 2000,
  end = as.numeric(format(Sys.time(), "\%Y")) - 2,
  vars = "tmax,tmin,dayl,prcp,srad,swe,vp",
  use_this_year = F
)
}
\arguments{
\item{site:}{site name (not important)}

\item{lat:}{latitude}

\item{lon:}{longitude}

\item{start:}{start of the range of years over which to download data}

\item{end:}{end of the range of years over which to download data}

\item{vars:}{string of variables to get for the site}

\item{use_last_year:}{set max year to last year, there may not be data for last year}
}
\value{
A data.frame of Daymet data for a given site
}
\description{
Get Daymet data for a given site and set of years
}
\examples{
get_daymet()
}
