% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_cdd_curves.R
\name{get_site_cdd_curves}
\alias{get_site_cdd_curves}
\title{Generate CDD curves for each site + year}
\usage{
get_site_cdd_curves(sites, ub, lb)
}
\arguments{
\item{sites:}{data.frame with siteid, latitude, longitude, year, and year_day columns}

\item{ub:}{the upper bound for degree day calculation in degrees Fahrenheit}

\item{lb:}{the lower bound for degree day calculation in degrees Fahrenheit}
}
\value{
A data.frame with cumulative degree days for the whole year for each site, truncated to include only the last day at 0 CDD and only the first day with the max CDD for the year + site.
}
\description{
Generate CDD curves for each site + year
}
