% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_siteyear_cdd.R
\name{get_siteyear_cdd}
\alias{get_siteyear_cdd}
\title{Get a cumulative degree day value for every day for a given site and year}
\usage{
get_siteyear_cdd(data, ub, lb)
}
\arguments{
\item{data:}{data.frame with latitude, longitude, year, and year_day columns}

\item{ub:}{the upper bound for degree day calculation in degrees Fahrenheit}

\item{lb:}{the lower bound for degree day calculation in degrees Fahrenheit}
}
\value{
A data.frame with cumulative degree days for the whole year for each site
}
\description{
Get a cumulative degree day value for every day for a given site and year
}
