% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cdd_density_tbl.R
\name{make_cdd_density_tbl}
\alias{make_cdd_density_tbl}
\title{Manually generate kernel density estimates for a range of CDD values, truncated to 0 CDD and the max CDD value in the input data.}
\usage{
make_cdd_density_tbl(data)
}
\arguments{
\item{data:}{data.frame with a cdd column}
}
\value{
A data.frame with a sequence of CDD values and density estimates, suitable for plotting with geom_line()
}
\description{
Manually generate kernel density estimates for a range of CDD values, truncated to 0 CDD and the max CDD value in the input data.
}
