% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_cdd_obs_site.R
\name{match_cdd_obs_site}
\alias{match_cdd_obs_site}
\title{Match observation CDDs to the closest value in a site of interest. The idea here is to say "given a bunch of observations from a range of locations, if I were to make that observation at a specific location, when would I make it?", assuming that CDD is interchangable between observed location and focal location.}
\usage{
match_cdd_obs_site(obs_data, site_curve)
}
\arguments{
\item{obs_data:}{data.frame with observations and CDD values}

\item{site_curve:}{data.frame with a full CDD curve for a given site of interest}
}
\value{
The obs_data data.frame with the closest CDD and date from the
}
\description{
Match observation CDDs to the closest value in a site of interest. The idea here is to say "given a bunch of observations from a range of locations, if I were to make that observation at a specific location, when would I make it?", assuming that CDD is interchangable between observed location and focal location.
}
