% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_listfiles.R
\name{s3_listfiles}
\alias{s3_listfiles}
\title{List names of files within a bucket on AWS S3}
\usage{
s3_listfiles(aws_profile, bucket_name, sub_dir = "")
}
\arguments{
\item{aws_profile}{Name of profile to access AWS. This must match .aws config file}

\item{bucket_name}{Name of AWS bucket}

\item{sub_dir}{Name of subdirectory within specified AWS bucket (optional)}
}
\value{
A character vector of paths to files in specified s3 bucket and sub directory
}
\description{
This function is a wrapper for functionality of paws list_objects. It parses list format returned by paws to return only file names in a given directory.
}
