% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ecodata.R
\name{theme_ecodata_dark}
\alias{theme_ecodata_dark}
\title{A light ggplot theme for EcoData Technology}
\usage{
theme_ecodata_dark(
  base_family = "Roboto Condensed",
  title_family = "Lexend Deca SemiBold",
  subtitle_family = "Lexend Deca Thin",
  axis_family = "Lexend Deca Light",
  base_color = "white",
  primary_color = "#1b2724",
  accent_color = "#6eb39c",
  gridlines = F,
  facet_outlines = F,
  base_size = 10
)
}
\arguments{
\item{base_family:}{the base font family}

\item{title_family:}{font to be used for the plot title}

\item{subtitle_family:}{font to be used for the subtitle}

\item{axis_family:}{font to be used for the axis titles}

\item{base_color:}{base color for text and lines}

\item{primary_color:}{color for primary components, including background}

\item{accent_color:}{color for accent components}

\item{base_size:}{base font size}

\item{gridlines:}{should gridlines be displayed?}

\item{facet_outlines:}{should facets be outlined?}
}
\value{
A ggplot theme
}
\description{
A light ggplot theme for EcoData Technology
}
\examples{
mtcars \%>\% ggplot(aes(x = wt, y = mpg)) + geom_point() + theme_ecodata_dark()
}
